+++ test/config/app.rb
@@ -7,3 +7,9 @@ end
 lowlevel_error_handler do |err|
   [200, {}, ["error page"]]
 end
+
+force_shutdown_error_response(
+  500,
+  {"Content-Type" => "application/json"},
+  ["{}"]
+)